[Feat]: ProjectPlanView 레이아웃 수정 #35
